更新restock的進貨數量可為浮點數

+++ pos_client/lib/view/sell/cashier.dart
@@ -385,7 +385,7 @@ class _CashierState extends State<Cashier> {
                   const Text('數量'),
                   const SizedBox(width: 8),
                   NumberInputWithIncrementDecrement(
-                    initialNumber: int.parse(quantity.text),
+                    initialNumber: double.parse(quantity.text),
                     onChanged: (value) {
                       quantity.text = value.toString();
                     },
